Imprementacion de presisicon

+++ interprete.h
@@ -21,9 +21,9 @@ class InterpreteExcepcion {
 //bool leer_flotante(std::string linea, int* indice, double* valor);
 
 typedef struct{ //Los ejes de movimiento deben de estar al principio de la estrucutra
-    double x; //eje x
-    double y; //eje y
-    double z; //eje z
+    long x; //eje x
+    long y; //eje y
+    long z; //eje z
     int N; // numero de linea
     double f; // feedrate
     int bandera_palabras;
